requirements.txt
dspy-ai>=2.4.0
pandas>=1.5.0
numpy>=1.21.0
tqdm>=4.64.0
requests>=2.28.0